메인 화면 씬 보완 1
- 엔터 눌러도 이제 창 안 없어짐
- 로그인 에러 별 메시지 창 출력
- 한글 출력 되는 임시 폰트로 변경(메이플 서체ㅎ)
- 기타 파일 정리 약간
