+++ Assets/Imported/Fading.unitypackage.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: da0f7a8396ab4644d9a6a33ca626f0c4
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
