+++ Assets/Imported/TextMesh Pro/Fonts/Maplestory Light SDF.asset.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 42fd96001e4175f44af9ca17f2772b7a
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 11400000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
