+++ Assets/Imported/TextMesh Pro/Fonts/Maplestory Light.ttf.meta
@@ -1,0 +1,22 @@
+fileFormatVersion: 2
+guid: b7ea6e4a7b3127148827b502f906ffb5
+TrueTypeFontImporter:
+  externalObjects: {}
+  serializedVersion: 4
+  fontSize: 16
+  forceTextureCase: -2
+  characterSpacing: 0
+  characterPadding: 1
+  includeFontData: 1
+  fontNames:
+  - Maplestory
+  fallbackFontReferences:
+  - {fileID: 12800000, guid: b0b34aab1fae5204da9f96b0cfb4027b, type: 3}
+  customCharacters: 
+  fontRenderingMode: 0
+  ascentCalculationMode: 1
+  useLegacyBoundsCalculation: 0
+  shouldRoundAdvanceValue: 1
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
