+++ Assets/Imported/npc1.unitypackage.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: 0d61d3a523b88af4ab91ff4e0e40ebce
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
