--- Assets/NpcMove.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: c97b08a6e2c719f4b9e032d9a5b084bd
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
