+++ Assets/Scripts/ShowKeyboard.cs
@@ -1,0 +1,41 @@
+using System.Collections;
+using System.Collections.Generic;
+using UnityEngine;
+using TMPro;
+using Microsoft.MixedReality.Toolkit.Experimental.UI;
+
+public class ShowKeyboard : MonoBehaviour
+{
+    private TMP_InputField inputField;
+
+    // Start is called before the first frame update
+    void Start()
+    {
+        inputField = GetComponent<TMP_InputField>();
+        inputField.onSelect.AddListener(x => OpenKeyboard());
+    }
+
+    public void OpenKeyboard()
+    {
+        NonNativeKeyboard.Instance.InputField = inputField;
+        NonNativeKeyboard.Instance.PresentKeyboard(inputField.text);
+
+        SetCaretColorAlpha(1);
+
+        NonNativeKeyboard.Instance.OnClosed += Instance_OnClosed;
+    }
+
+    private void Instance_OnClosed(object sender, System.EventArgs e)
+    {
+        SetCaretColorAlpha(0);
+        NonNativeKeyboard.Instance.OnClosed -= Instance_OnClosed;
+    }
+
+    public void SetCaretColorAlpha(float value)
+    {
+        inputField.customCaretColor = true;
+        Color caretColor = inputField.caretColor;
+        caretColor.a = value;
+        inputField.caretColor = caretColor;
+    }
+}
