--- Assets/ShowKeyboard.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: be9b87f513477034fb06ed10c3dc636c
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
